Huge simplification refactor: remove "root sub()" semantics
Since the Store is already a properly-typed `IRef`, why not just deref()
that directly? This significantly simplifies a lot of code and removes a
bunch of unnecessary implementation cruft.

Now the downside is of course that subscriptions are now tightly coupled
to a single Store instance. However, given that Store can `loadSnapshot`
and we can use `dispatchSync` on it in tests, this doesn't seem too
terrible, and the pros outweigh this con:

- Subscriptions can now be `deref()`'d in arbitrary code, which is a
huge benefit not just for testing but for real world applications
- No need to export the type of your store, just the instance (which you
already had to anyway for the Provider). This is minor, but having to
create the root sub with `sub<IStoreType>()` seems like unnecessary
cognitive load. `store.deref()` seems very intuitive to get the full state
- Less code, simpler code. This change makes The Mall more lightweight,
with less moving parts to maintain.

This change also makes sure that one-off Reference uses (eg:
`mySub().deref()`) don't stay cached, leaking memory. Since we now
support arbitrary `deref()`, this was fairly important to include in
this commit.